+++ core/src/it/units/sdm/project/board/gui/GuiFreedomBoard.java
@@ -10,7 +10,7 @@ import com.badlogic.gdx.scenes.scene2d.ui.*;
 import it.units.sdm.project.board.Board;
 import it.units.sdm.project.board.Position;
 import it.units.sdm.project.exceptions.InvalidPositionException;
-import it.units.sdm.project.game.GameStatusHandler;
+import it.units.sdm.project.game.gui.GameStatusHandler;
 import it.units.sdm.project.game.gui.FreedomGame;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
