+++ core/src/it/units/sdm/project/game/gui/screens/MainMenuScreen.java
@@ -1,4 +1,4 @@
-package it.units.sdm.project.game.gui;
+package it.units.sdm.project.game.gui.screens;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.Screen;
@@ -9,6 +9,7 @@ import com.badlogic.gdx.scenes.scene2d.Stage;
 import com.badlogic.gdx.scenes.scene2d.ui.*;
 import com.badlogic.gdx.utils.ScreenUtils;
 import com.badlogic.gdx.utils.viewport.FitViewport;
+import it.units.sdm.project.game.gui.FreedomGame;
 import org.jetbrains.annotations.NotNull;
 
 public class MainMenuScreen implements Screen {
